refactor: move format-fee-tier-options from send-bitcoin hooks to shared utils
